backend/requirements/development.txt
-r production.txt

mypy==0.920
flake8==4.0.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-aiohttp==0.3.0
black==21.12b0
py-spy==0.3.11

aiohttp-devtools==0.13.1
aiohttp-debugtoolbar==0.6.0
